Fixed several bugs, including two bugs related to errors in the paper.
There are several bugs in the paper (Igarashi, Takeo, and Yuki
Igarashi. "Implementing as-rigid-as-possible shape manipulation and surface
flattening." Journal of Graphics, GPU, and Game Tools 14.1 (2009): 17-30.),
and two of them are addressed in this commit:

1. The $T_k'$ normalization factor on page 25 is incorrect.  It should read
$T_k' = \frac{1}{\sqrt{c_k^2+s_k^2}}.

2. The formulation of $G_k$ on page 23 is incorrect, as it appears to be
rotation-variant.

(While we are on the subject, there is a transpose marker missing on several
instances of the formula $(G^tG)^{-1}G^t$.)

Also, I had implemented the following incorrectly, and I would like to point
it out since I think it could be a stumbling block for anyone else
implementing this paper.  The edge vectors $$ on page 26 should come from the
*untransformed* mesh not the *similarity-transformed* one (which probably
would have been written with a prime marker, e.g. "$e_0'$").